feat: auths 모듈 정리 & alarms 엔티티 추가 및 관련 CRUD 생성

+++ test/work/work.e2e-spec.ts
@@ -181,10 +181,10 @@ describe('UsersController (e2e)', () => {
         .get(`/work/today`)
         .set('Cookie', testAccessToken.admin);
 
-      await workRepository.query(`delete from "work"`);
+      await workRepository.query(`delete from "work" where "userUuid" in ('${testIds.admin}', '${testIds.manager}', '${testIds.user}')`);
 
       expect(response.status).toBe(200);
-      expect(response.body.length).toBe(3);
+      expect(response.body.length).toBeGreaterThanOrEqual(3);
     });
   })
